Move bounds methods to a separate module
This allows them to be reused in the battery pool, without leading to
circular imports.
